+++ Trigger/src/FilterEcalMVATrigger_module.cc
@@ -170,7 +170,7 @@ namespace mu2e {
     if (_step==4) return false;
 
     for (CaloTrigSeedCollection::const_iterator seedIt = caloTrigSeeds.begin(); seedIt != caloTrigSeeds.end(); ++seedIt){
-      disk= cal.crystal(seedIt->crystalid()).diskId();
+      disk= cal.crystal((int)seedIt->crystalid()).diskId();
       _fdiskpeak   = (float) disk;
       _Epeak   = seedIt->epeak();
       _tpeak   = seedIt->tpeak()+_TOFF;
